// src/core/utility/enums/core.enum.ts
enum PftRoutes {
  LOGIN = '/login',
  SIGNUP = '/sign-up',
  PENDING_REQUESTS = 'pending-requests',
  APPROVED_REQUESTS = 'approved-requests',
  REJECTED_REQUESTS = 'rejected-requests',
  TRANSACTIONS = 'transactions',
  DASHBOARD = 'dashboard',
  INVESTMENTS = 'investments',
  INCOME = 'income',
  EXPENSE = 'expense',
  USER_REPORTS = 'user-reports',
  PAGE_NOT_FOUND = '*',
  CALENDAR = 'calendar',
}

export { PftRoutes };